pyproject.toml
[project]
name = "kiwoom-autotrading"
version = "0.1.0"
description = "키움증권 REST API 기반 모의투자 자동매매 시스템"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "apscheduler>=3.10.0",
    "python-telegram-bot>=21.0",
    "structlog>=24.0.0",
    "ta>=0.11.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx",  # for TestClient
    "ruff>=0.5.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: 통합 테스트 (모의투자 API 필요)",
    "market_hours: 장중에만 실행 가능한 테스트",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
